4. TESTE DE API/test_api_python/back-end/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
PyMySQL==1.0.3
python-dotenv==1.0.0